requirements.txt
streamlit
pandas
Pillow
streamlit-autorefresh
